+++ clap/__init__.py
@@ -3,5 +3,6 @@ from clap import errors
 from clap import option
 from clap import parser
 from clap import modes
+from clap import base
 
 __version__ = '0.6.3'